+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Wait for initial cache sync before starting the server.
+
 ## [1.1.3] - 2023-03-01
 
 ### Changed